fix(chat): `ChatCompletion.created` field change type from `Int` to `Long` (#282)

+++ openai-core/src/commonMain/kotlin/com.aallam.openai.api/chat/ChatCompletion.kt
@@ -20,7 +20,7 @@ public data class ChatCompletion(
     /**
      * The creation time in epoch milliseconds.
      */
-    @SerialName("created") public val created: Int,
+    @SerialName("created") public val created: Long,
 
     /**
      * The model used.
